[Refactor] Add date utils
